Trim datCalc d4e0 float temporaries

+++ src/Graphics/Model/mdlFile.c
@@ -320,9 +320,9 @@ u32 FUN_00322fb0(u32 param_1);
 void FUN_00322fd0(int param_1,u32 param_2,u32 *param_3);
 u32 FUN_003230f0(u32 param_1,u32 param_2,u32 param_3);
 int FUN_00323240(u32 param_1,int param_2,u16 param_3,int param_4);
-int FUN_003233a0(u32 param_1,u32 param_2,u16 param_3,int param_4,u64 param_5);
+int FUN_003233a0(u32 param_1,u32 param_2,u16 param_3,int param_4,u32 param_5);
 #pragma alias FUN_003233a0_ptr FUN_003233a0
-extern int FUN_003233a0_ptr(u32 param_1,u32 param_2,u16 param_3,u8 *param_4,u64 param_5);
+extern int FUN_003233a0_ptr(u32 param_1,u32 param_2,u16 param_3,u8 *param_4,u32 param_5);
 u64 FUN_003234f0(u16 param_1,u32 param_2);
 #pragma alias FUN_003234f0_u32 FUN_003234f0
 extern u32 FUN_003234f0_u32(u16 param_1,u32 param_2);
@@ -7050,7 +7050,7 @@ int FUN_00323240(u32 param_1,int param_2,u16 param_3,int param_4)
 // FUN_003233A0 NONMATCHING
 
 
-int FUN_003233a0(u32 param_1,u32 param_2,u16 param_3,int param_4,u64 param_5)
+int FUN_003233a0(u32 param_1,u32 param_2,u16 param_3,int param_4,u32 param_5)
